Basic refactor for MediaType
