+++ cosmos-common/src/main/scala/com/mesosphere/cosmos/circe/Decoders.scala
@@ -4,7 +4,7 @@ import cats.syntax.either._
 import com.mesosphere.cosmos.error.JsonDecodingError
 import com.mesosphere.cosmos.error.JsonParsingError
 import com.mesosphere.cosmos.finch.MediaTypedDecoder
-import com.mesosphere.cosmos.http.MediaType
+import com.mesosphere.http.MediaType
 import com.netaporter.uri.Uri
 import io.circe.Decoder
 import io.circe.DecodingFailure
