+++ cosmos-common/src/main/scala/com/mesosphere/cosmos/error/UnsupportedContentType.scala
@@ -1,6 +1,6 @@
 package com.mesosphere.cosmos.error
 
-import com.mesosphere.cosmos.http.MediaType
+import com.mesosphere.http.MediaType
 import io.circe.Encoder
 import io.circe.JsonObject
 import io.circe.generic.semiauto.deriveEncoder
