+++ cosmos-common/src/main/scala/com/mesosphere/cosmos/finch/FinchExtensions.scala
@@ -1,9 +1,8 @@
 package com.mesosphere.cosmos.finch
 
-import com.mesosphere.cosmos.http.CompoundMediaType
-import com.mesosphere.cosmos.http.CompoundMediaTypeParser
-import com.mesosphere.cosmos.http.MediaType
-import com.mesosphere.cosmos.http.MediaTypeOps
+import com.mesosphere.http.CompoundMediaType
+import com.mesosphere.http.CompoundMediaTypeParser
+import com.mesosphere.http.MediaType
 import io.circe.Json
 import io.finch.DecodeEntity
 import io.finch.Endpoint
@@ -14,7 +13,7 @@ object FinchExtensions {
 
   def beTheExpectedTypes(expectedTypes: List[MediaType]): ValidationRule[MediaType] =
     ValidationRule(s"match one of ${expectedTypes.map(_.show).mkString(", ")}") { actual =>
-      expectedTypes.exists(expected => MediaTypeOps.compatible(expected, actual))
+      expectedTypes.exists(expected => MediaType.compatible(expected, actual))
     }
 
   implicit val decodeMediaType: DecodeEntity[MediaType] = {
