+++ cosmos-common/src/main/scala/com/mesosphere/cosmos/http/RequestSession.scala
@@ -1,5 +1,7 @@
 package com.mesosphere.cosmos.http
 
+import com.mesosphere.http.MediaType
+
 case class RequestSession(
   authorization: Option[Authorization],
   originInfo: OriginHostScheme,
