+++ cosmos-common/src/main/scala/com/mesosphere/universe/PackageDefinition.scala
@@ -4,8 +4,8 @@ import cats.data.NonEmptyList
 import com.mesosphere.cosmos.circe.Decoders._
 import com.mesosphere.cosmos.finch.MediaTypedDecoder
 import com.mesosphere.cosmos.finch.MediaTypedEncoder
-import com.mesosphere.cosmos.http.MediaType
 import com.mesosphere.cosmos.rpc.v1.model.PackageCoordinate
+import com.mesosphere.http.MediaType
 import com.mesosphere.universe
 import com.mesosphere.universe
 import io.circe.Decoder
