+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/handler/RequestErrorsSpec.scala
@@ -1,11 +1,11 @@
 package com.mesosphere.cosmos.handler
 
-import com.mesosphere.cosmos.http.CompoundMediaTypeParser
 import com.mesosphere.cosmos.http.HttpRequest
 import com.mesosphere.cosmos.http.PackageRpcPath
 import com.mesosphere.cosmos.rpc.MediaTypes
 import com.mesosphere.cosmos.rpc.v1.model.PackageRepositoryAddRequest
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient._
+import com.mesosphere.http.CompoundMediaTypeParser
 import com.netaporter.uri.dsl._
 import com.twitter.finagle.http.Fields
 import com.twitter.finagle.http.Status
