+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/HttpClient.scala
@@ -2,21 +2,21 @@ package com.mesosphere.cosmos
 
 import com.mesosphere.cosmos.error.UnsupportedContentEncoding
 import com.mesosphere.cosmos.error.UnsupportedRedirect
-import com.mesosphere.cosmos.http.MediaType
-import com.mesosphere.cosmos.http.MediaTypeParser
+import com.mesosphere.http.MediaType
+import com.mesosphere.http.MediaTypeParser
 import com.netaporter.uri.Uri
 import com.twitter.finagle.http.Fields
+import com.twitter.finagle.http.filter.LogFormatter
 import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.util.Future
 import com.twitter.util.Return
 import com.twitter.util.Throw
+import com.twitter.util.Time
 import java.io.IOException
 import java.io.InputStream
 import java.net.HttpURLConnection
 import java.net.MalformedURLException
 import java.net.URISyntaxException
-import com.twitter.finagle.http.filter.LogFormatter
-import com.twitter.util.Time
 import java.util.TimeZone
 import java.util.zip.GZIPInputStream
 import org.apache.commons.lang.time.FastDateFormat
