+++ cosmos-test-common/src/main/scala/com/mesosphere/Generators.scala
@@ -1,6 +1,6 @@
 package com.mesosphere
 
-import com.mesosphere.cosmos.http.MediaType
+import com.mesosphere.http.MediaType
 import com.netaporter.uri.Uri
 import io.circe.testing.instances._
 import java.nio.ByteBuffer
