+++ cosmos-test-common/src/test/scala/com/mesosphere/cosmos/finch/DispatchingMediaTypedEncoderSpec.scala
@@ -1,8 +1,8 @@
 package com.mesosphere.cosmos.finch
 
 import cats.data.NonEmptyList
-import com.mesosphere.cosmos.http.CompoundMediaType
-import com.mesosphere.cosmos.http.MediaType
+import com.mesosphere.http.CompoundMediaType
+import com.mesosphere.http.MediaType
 import io.circe.Encoder
 import io.circe.Json
 import io.circe.syntax._
