+++ cosmos-test-common/src/test/scala/com/mesosphere/cosmos/finch/RequestValidatorsSpec.scala
@@ -3,16 +3,16 @@ package com.mesosphere.cosmos.finch
 import cats.data.NonEmptyList
 import com.mesosphere.Generators.Implicits._
 import com.mesosphere.cosmos.http.Authorization
-import com.mesosphere.cosmos.http.CompoundMediaTypeParser
 import com.mesosphere.cosmos.http.Get
 import com.mesosphere.cosmos.http.HttpRequest
 import com.mesosphere.cosmos.http.HttpRequestMethod
-import com.mesosphere.cosmos.http.MediaType
 import com.mesosphere.cosmos.http.Post
 import com.mesosphere.cosmos.http.RawRpcPath
 import com.mesosphere.cosmos.http.RequestSession
 import com.mesosphere.cosmos.httpInterface
 import com.mesosphere.cosmos.rpc.MediaTypes._
+import com.mesosphere.http.CompoundMediaTypeParser
+import com.mesosphere.http.MediaType
 import com.mesosphere.util.urlSchemeHeader
 import com.twitter.finagle.http.Fields
 import com.twitter.finagle.http.Fields
