+++ cosmos-test-common/src/test/scala/com/mesosphere/cosmos/finch/VersionedResponsesSpec.scala
@@ -3,8 +3,8 @@ package com.mesosphere.cosmos.finch
 import com.mesosphere.cosmos.finch.TestingMediaTypes._
 import com.mesosphere.cosmos.http.HttpRequest
 import com.mesosphere.cosmos.http.RawRpcPath
-import com.mesosphere.cosmos.http.MediaType
 import com.mesosphere.cosmos.http.RequestSession
+import com.mesosphere.http.MediaType
 import com.twitter.io.Buf
 import com.twitter.util.Future
 import com.twitter.util.Try
